% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Metric computation.R
\name{add.met}
\alias{add.met}
\title{Reference specific Benthic Metric Calculation}
\usage{
add.met(Test, Reference)
}
\arguments{
\item{Test.taxa}{Output of \code{\link[BenthicAnalysis]{benth.met}}$Raw.Data for the test site}

\item{Reference.taxa}{Output of \code{\link[BenthicAnalysis]{benth.met}}$Raw.Data for the Reference site}
}
\value{
Indicator metrics caluclated from \code{\link[BenthicAnalysis]{benth.met}} in addition to O:E ratios and Bray-Curtis Distance.
}
\description{
Calculation of a variety of metrics for determing impairment of Benthic Macroinvertebrate Communities, based on available Reference Condition information.
}
\examples{
#load datasets
data(YKEnvData,envir = environment()) #Biological dataset
data(YKBioData,envir = environment()) #Environmental dataset

#Calculate indicator metrics from raw biological data
bio.data.test<-benth.met(YKBioData,2,2)

#Extract just the summary metrics
bio.data<-bio.data.test$Summary.Metrics

#standardize row names between datasets
rownames(YKEnvData)<-bio.data.test$Site.List

#Match a test site (#201) to the nearest neighbour reference set
nn.sites<-site.match(YKEnvData[201,-c(1)],YKEnvData[1:118,-c(1)],k=NULL,adaptive=T)

#Calculate additional metrics based on selected Reference sites
add.met(Test=bio.data.test$Raw.Data[201,],Reference=bio.data.test$Raw.Data[names(nn.sites$final.dist),])
}
\keyword{Benthic}
\keyword{Metrics}

