% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting Functions.R
\name{boxplot.tsa.object}
\alias{boxplot.tsa.object}
\title{z-score Boxplots}
\usage{
\method{boxplot}{tsa.object}(tsa.object)
}
\arguments{
\item{tsa.object}{An object from tsa.test()}
}
\description{
Boxplots of z-scores used in Test Site Analysis. Metrics used in calculation of TSA are circled, metrics that
contribute significantly to an impairment score are marked with an * along the x axis.
}
\examples{
#load datasets
data(YKEnvData,envir = environment()) #Biological dataset
data(YKBioData,envir = environment()) #Environmental dataset

#Calculate indicator metrics from raw biological data
bio.data.test<-benth.met(YKBioData,2,2)

#Extract just the summary metrics
bio.data<-bio.data.test$Summary.Metrics

#standardize row names between datasets
rownames(YKEnvData)<-bio.data.test$Site.List

#Match a test site (#201) to the nearest neighbour reference set
nn.sites<-site.match(YKEnvData[201,-c(1)],YKEnvData[1:118,-c(1)],k=F,adaptive=T)

#Extract the raw taxa data for additional metric calculation
taxa.data<-rbind(bio.data.test$Raw.Data[names(nn.sites$final.dist),],bio.data.test$Raw.Data[rownames(bio.data[201,]),])

#TSA test of indicator metrics at test site and reference sites selected used site.match()
tsa.results<-tsa.test(bio.data[201,],bio.data[names(nn.sites$final.dist),],distance=nn.sites$final.dist, outlier.rem=T, m.select=T, add.metric=taxa.data)

Evaluate Results
boxplot(tsa.results)
}
\keyword{Analysis,}
\keyword{Boxplot}
\keyword{Site}
\keyword{Test}

