% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting Functions.R
\name{sitematch.plot}
\alias{sitematch.plot}
\title{Scatterplot of ANNA and RDA-ANNA site matching functions}
\usage{
sitematch.plot(match.object, axis = c(1, 2))
}
\arguments{
\item{axis}{A two character vector indicating which axis to plot}

\item{site.match.object}{An object from \code{\link[BenthicAnalysis]{"site.match"}}}
}
\description{
Scatterplot of ANNA and RDA-ANNA site matching functions.
}
\examples{
#load datasets
data(YKEnvData,envir = environment()) #Biological dataset
data(YKBioData,envir = environment()) #Environmental dataset

#Calculate indicator metrics from raw biological data
bio.data.test<-benth.met(bio.data.yk1,2,2)

#Extract just the summary metrics
bio.data<-bio.data.test$Summary.Metrics

#standardize row names between datasets
rownames(env.data.yk)<-bio.data.test$Site.List

#Match a test site (#201) to the nearest neighbour reference set
nn.sites<-site.match(env.data.yk[201,-c(1)],env.data.yk[1:118,-c(1)],k=F,adaptive=T)

#Extract the raw taxa data for additional metric calculation
taxa.data<-rbind(bio.data.test$Raw.Data[names(nn.sites$final.dist),],bio.data.test$Raw.Data[rownames(bio.data[201,]),])

#TSA test of indicator metrics at test site and reference sites selected used site.match()
tsa.results<-tsa.test(bio.data[201,],bio.data[names(nn.sites$final.dist),],distance=nn.sites$final.dist, outlier.rem=T, m.select=T, add.metric=taxa.data)

Evaluate Results
sitematch.plot(nn.sites)
}
\keyword{ANNA,}
\keyword{RDA-ANNA}

