% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{tsa.zscore}
\alias{tsa.zscore}
\title{Z-Score Calculation}
\usage{
tsa.zscore(Test, Reference)
}
\arguments{
\item{Test}{Data frame of metric scores at the test site.}

\item{Reference}{Data frame of metric scores at the reference sites}
}
\description{
This calculates z-scores centered on Reference Sites
}
\examples{
tsa.zscore()
}
\keyword{z-score}

